Rework email() method
The email() method now takes an EmailUri rather than a String containing
an email address. The new email_from_address() method can be used to look
up an Email object from an email address.

This change is done for consistency, and because Document objects return
the shepherd as an Option<EmailUri> so we need a method to retrieve via
EmailUri.